==> scripts/2-analysis_temporal&spatial.R <==
################################################################################
##
## Script for computing taxonomic and functional diversity between years for kelp sites,
##
## and also for spatial data
## 
## Code by Camille Magneville, Sebastien Villeger and Paula Sgarlatta
##
################################################################################

rm(list=ls()) # cleaning memory

# libraries
library(tidyverse)
library(here)
library(mFD)


# loading data
load(here::here("data", "kelp_sp_maxN.RData") )
load(here::here("data", "sp_3D_coord.RData") ) 
load(here::here("data", "spatial_sp_biom.RData") )
biomass <- read.csv(here::here("data", "spatial_biomass.csv"))


## computing taxonomic and functional diversity for kelp sites ####

temporal_fd_kelp <- mFD::alpha.fd.multidim(
  sp_faxes_coord   = sp_3D_coord,
  asb_sp_w         = kelp_sp_maxN,
  ind_vect         = c("fide", "fric", "fdis"),
  scaling          = TRUE,
  check_input      = TRUE,
  details_returned = TRUE)

temporal_alpha_kelp <- temporal_fd_kelp$functional_diversity_indices

## computing taxonomic and functional diversity for spatial data ####

spatial_fd <- mFD::alpha.fd.multidim(
  sp_faxes_coord   = sp_3D_coord,
  asb_sp_w         = spatial_sp_biom,
  ind_vect         = c("fide", "fric", "fdis"),
  scaling          = TRUE,
  check_input      = TRUE,
  details_returned = TRUE)

spatial_alpha <- spatial_fd$functional_diversity_indices

## Adding biomass for statistics

spatial <- spatial_alpha %>% 
  rownames_to_column("Site") %>% 
  left_join(biomass, by="Site")

# saving ####

save(temporal_fd_kelp, file=here::here("data", "temporal_fd_kelp_biomass.RData") )
save(temporal_alpha_kelp, file=here::here("data",  "temporal_alpha_kelp_biomass.RData") )

save(spatial_fd, file=here::here("data",  "spatial_fd_biomass.RData") )
save(spatial, file=here::here("data", "spatial_alpha_biomass.RData") )


#################################### end of script ####################################################################
